OnboardManagement/Onboarding/requirements.txt
aniso8601==8.0.0
bcrypt==3.2.0
boto3==1.14.44
botocore==1.17.44
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-mongoengine==0.9.5
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mongoengine==0.20.0
pycparser==2.20
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
uWSGI==2.0.19.1
Werkzeug==1.0.1
WTForms==2.3.3
